+++ HISTORY.md
@@ -1,3 +1,8 @@
+## 0.1.3 (2010-09-10)
+
+* Depend on rack/mime instead of mime-types
+* Add Content-Length header
+
 ## 0.1.2 (2010-09-09)
 
 * Add a way to customize mapping of PATH_INFO to MogileFS keys